+++ lib/cosmo/api.rb
@@ -5,6 +5,7 @@ require "cosmo/api/counter"
 require "cosmo/api/kv"
 require "cosmo/api/stats"
 require "cosmo/api/cron"
+require "cosmo/api/batch"
 
 module Cosmo
   module API