+++ spec/support/global_helpers.rb
@@ -16,6 +16,8 @@ RSpec.shared_context "Global helpers" do
     Cosmo::API::Counter.instance_variable_set(:@instance, nil)
     Cosmo::Publisher.instance_variable_set(:@instance, nil)
     Cosmo::Job::Limit.instance_variable_set(:@instance, nil)
+    Cosmo::Batch.instance_variable_set(:@counter, nil)
+    Cosmo::Batch.instance_variable_set(:@kv, nil)
   end
 
   def wait_until(timeout:)